+++ app/src/main/java/com/jraska/github/client/AppComponent.kt
@@ -42,7 +42,6 @@ interface AppComponent {
 
 @Module(
   includes = [
-    AppModule::class,
     ConfigDebugModule::class,
     CoreAndroidModule::class,
     ChromeCustomTabsModule::class,
