+++ core/build.gradle
@@ -27,6 +27,8 @@ dependencies {
   implementation 'androidx.lifecycle:lifecycle-extensions:2.2.0'
   implementation 'org.threeten:threetenbp:1.4.4:no-tzdb'
 
+  implementation 'com.jakewharton.threetenabp:threetenabp:1.3.0'
+  implementation rootProject.ext.fresco
 
   kapt rootProject.ext.daggerAnnotationProcessor
   implementation rootProject.ext.dagger
